+++ lib/src/code_generator/objc_nullable.dart
@@ -45,6 +45,9 @@ class ObjCNullable extends Type {
   bool get sameDartAndCType => false;
 
   @override
+  bool get sameDartAndFfiDartType => false;
+
+  @override
   String convertDartTypeToFfiDartType(
     Writer w,
     String value, {