+++ lib/src/code_generator/pointer.dart
@@ -88,6 +88,9 @@ class ObjCObjectPointer extends PointerType {
   bool get sameDartAndCType => false;
 
   @override
+  bool get sameDartAndFfiDartType => false;
+
+  @override
   String convertDartTypeToFfiDartType(
     Writer w,
     String value, {