+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - When generating typedefs for `Pointer<NativeFunction<Function>>`, also
   generate a typedef for the `Function`.
 - Use Dart wrapper types in args and returns of ObjCBlocks.
+- Use Dart wrapper types in args and returns of static functions.
 - Bump min SDK version to 3.2.0-114.0.dev.
 - Renamed `asset` to `assetId` for `ffi-native`.